requirements.txt
absl-py==1.1.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
backports.ssl-match-hostname==3.7.0.1
beautifulsoup4==4.11.1
bert==2.2.0
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
dash==2.5.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
erlastic==2.0.0
executing==0.8.3
Flask==2.1.2
Flask-Compress==1.12
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gevent==21.12.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
interpret==0.2.7
interpret-core==0.2.7
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.2
kiwisolver==1.4.2
kt-legacy==1.0.4
libclang==14.0.1
lime==0.2.0.1
llvmlite==0.38.1
mail-parser==3.15.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multiprocess==0.70.13
nest-asyncio==1.5.5
networkx==2.8.4
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathos==0.2.9
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.9.0
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyspellchecker==0.6.3
python-dateutil==2.8.2
python-whois==0.7.3
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
pyzmq==23.2.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
SALib==1.4.5
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
shap==0.41.0
simplejson==3.17.6
six==1.16.0
skope-rules==1.0.1
slicer==0.0.7
soupsieve==2.3.2.post1
stack-data==0.3.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
treeinterpreter==0.2.3
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.20.0
whois== 0.9.27
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
